requirements.txt
Django==4.2
djangorestframework
mysqlclient
python-dotenv
djangorestframework-simplejwt
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
django-extensions

